requirements.txt
Django==2.2
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework-gis==0.14
psycopg2==2.8.1
pytz==2019.1
sqlparse==0.3.0
